# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Place {
  id: string;
  name: string;
  sports: string[];
  distance: string;
  address: string;
  latitude: number;
  longitude: number;
  accent: string;
}

export interface Game {
  id: string;
  /** @nullable */
  placeId: string | null;
  placeName: string;
  sport: string;
  title: string;
  host: string;
  players: number;
  playerLimit: number;
  time: string;
  date: string;
  noFixedTime: boolean;
  skillLevel: string;
  joined: boolean;
}

export type PlaceDetails = Place & {
  games: Game[];
};

export interface GameInput {
  /** @nullable */
  placeId: string | null;
  placeName: string;
  sport: string;
  title: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  playerLimit: number;
  date: string;
  time: string;
  noFixedTime: boolean;
  skillLevel: string;
}

export interface InterestGroup {
  id: string;
  sport: string;
  name: string;
  description: string;
  members: number;
  memberLimit: number;
  location: string;
  timing: string;
  joined: boolean;
  host: string;
}

export interface GroupInput {
  sport: string;
  name: string;
  description: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  memberLimit: number;
  location?: string;
  timing?: string;
}

export interface Player {
  id: string;
  name: string;
  avatar: string;
  city: string;
  sports: string[];
  level: string;
  verified: boolean;
  bio: string;
  achievements: string[];
  videos: string[];
  lookingFor: string;
}

export interface ProfileUpdate {
  name?: string;
  avatar?: string;
  city?: string;
  bio?: string;
  sports?: string[];
  level?: string;
  lookingFor?: string;
  videos?: string[];
}

export interface SkillVideoInput {
  filename: string;
  contentType: string;
  /** Inline video data */
  videoBase64: string;
}

export interface SkillAnalysis {
  summary: string;
  strengths: string[];
  focus: string[];
  confidence: string;
}