# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: string;
  /** @nullable */
  placeId: string | null;
  placeName: string;
  sport: string;
  title: string;
  host: string;
  players: number;
  playerLimit: number;
  time: string;
  date: string;
  noFixedTime: boolean;
  skillLevel: string;
  joined: boolean;
}