# lib/api-zod/src/generated/types/gameInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameInput {
  /** @nullable */
  placeId: string | null;
  placeName: string;
  sport: string;
  title: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  playerLimit: number;
  date: string;
  time: string;
  noFixedTime: boolean;
  skillLevel: string;
}