# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  sport: string;
  name: string;
  description: string;
  /**
     * @minimum 2
     * @maximum 50
     */
  memberLimit: number;
  location?: string;
  timing?: string;
}