# lib/api-zod/src/generated/types/interestGroup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterestGroup {
  id: string;
  sport: string;
  name: string;
  description: string;
  members: number;
  memberLimit: number;
  location: string;
  timing: string;
  joined: boolean;
  host: string;
}