# lib/api-zod/src/generated/types/place.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Place {
  id: string;
  name: string;
  sports: string[];
  distance: string;
  address: string;
  latitude: number;
  longitude: number;
  accent: string;
}