# lib/api-zod/src/generated/types/placeDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from './game';
import type { Place } from './place';

export type PlaceDetails = Place & {
  games: Game[];
};