# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: string;
  name: string;
  avatar: string;
  city: string;
  sports: string[];
  level: string;
  verified: boolean;
  bio: string;
  achievements: string[];
  videos: string[];
  lookingFor: string;
}