# lib/api-zod/src/generated/types/skillVideoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillVideoInput {
  filename: string;
  contentType: string;
  /** Inline video data */
  videoBase64: string;
}